consumer slf4j loading, main object wrapping
