+++ consumer/build.gradle.kts
@@ -13,7 +13,7 @@ java {
 }
 
 application {
-    mainClass.set("consumer.MainKt")
+    mainClass.set("consumer.ConsumerAppMain")
 }
 
 dependencies {
@@ -22,6 +22,8 @@ dependencies {
     implementation("org.springframework:spring-context")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3")
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
+
     implementation("org.slf4j:slf4j-api:1.7.36")
     implementation("ch.qos.logback:logback-classic:1.2.11")
+    implementation("ch.qos.logback:logback-core:1.2.11")
 }
